move styles out and update paths
