+++ src/components/Button/index.js
@@ -1,0 +1,5 @@
+import Button from './button'
+import { LinkButton, NakedButton, OutlinedButton } from './button.styles'
+
+export default Button
+export { LinkButton, NakedButton, OutlinedButton }
