+++ src/components/CallToAction/index.js
@@ -1,0 +1,3 @@
+import CallToAction from './callToAction'
+
+export default CallToAction
