+++ src/components/EventCard/index.js
@@ -1,0 +1,5 @@
+import EventCard from './eventCard'
+import { EventCardLink } from './eventCard.styles'
+
+export { EventCardLink }
+export default EventCard
