+++ src/components/Events/index.js
@@ -1,0 +1,3 @@
+import Events from './events'
+
+export default Events
