+++ src/components/Hero/hero.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { HeroWrapper, HeroText, Block } from './hero.styles'
+
+const Hero = () => (
+    <HeroWrapper>
+        <HeroText>
+            Front End Developers &amp; Designers with <Block>.class</Block>
+        </HeroText>
+    </HeroWrapper>
+)
+
+export default Hero
