+++ src/components/Hero/index.js
@@ -1,0 +1,3 @@
+import Hero from './hero'
+
+export default Hero
