+++ src/components/Icon/icon.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { IconWrapper } from './icon.styles'
+
+const Icon = ({ name }) => (
+  <IconWrapper name={name} />
+)
+
+export default Icon
