+++ src/components/Icon/index.js
@@ -1,0 +1,5 @@
+import Icon from './icon'
+import { IconWrapper } from './icon.styles'
+
+export default Icon
+export { IconWrapper }
