+++ src/components/Layout/layout.js
@@ -1,17 +1,13 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import Helmet from 'react-helmet'
-import styled from 'react-emotion'
 import { StaticQuery, graphql, withPrefix } from 'gatsby'
-import { COLOURS } from '../constants'
-
-import Header from '../components/Header'
-import appleTouch from '../assets/img/apple-touch-icon.png'
-import fav32 from '../assets/img/favicon-32x32.png'
-import fav16 from '../assets/img/favicon-16x16.png'
-import maskIcon from '../assets/img/safari-pinned-tab.svg'
-
-import '../assets/css/fenders-styles.css'
+import { PageContainer, Wrapper } from './layout.styles'
+import Header from '../Header'
+import appleTouch from '../../assets/img/apple-touch-icon.png'
+import fav32 from '../../assets/img/favicon-32x32.png'
+import fav16 from '../../assets/img/favicon-16x16.png'
+import maskIcon from '../../assets/img/safari-pinned-tab.svg'
 
 const Layout = ({ children }) => (
     <StaticQuery
@@ -69,14 +65,3 @@ Layout.propTypes = {
 }
 
 export default Layout
-
-const PageContainer = styled.div`
-`
-
-const Wrapper = styled.div`
-    max-width: 1800px;
-    margin: 0 auto;
-    width: 100%;
-    position: relative;
-    background: ${COLOURS.PURPLE_BG};
-`
