+++ src/components/Layout/layout.styles.js
@@ -1,0 +1,13 @@
+import styled from 'react-emotion'
+import { COLOURS } from '../../constants'
+
+export const PageContainer = styled.div`
+`
+
+export const Wrapper = styled.div`
+    max-width: 1800px;
+    margin: 0 auto;
+    width: 100%;
+    position: relative;
+    background: ${COLOURS.PURPLE_BG};
+`
